python_server/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.1.post1
joblib==1.3.2
tensorflow==2.15.0
flask-socketio==5.3.6
kafka-python==2.0.2
```
